+++ libraries/muxer/src/main/java/androidx/media3/muxer/FragmentedMp4Muxer.java
@@ -83,7 +83,7 @@ import java.nio.ByteBuffer;
  * </ul>
  */
 @UnstableApi
-public final class FragmentedMp4Muxer implements Muxer {
+public final class FragmentedMp4Muxer {
   /** The default fragment duration. */
   public static final long DEFAULT_FRAGMENT_DURATION_MS = 2_000;
 
@@ -178,7 +178,14 @@ public final class FragmentedMp4Muxer implements Muxer {
     trackIdToTrack = new SparseArray<>();
   }
 
-  @Override
+  /**
+   * Adds a track of the given media format.
+   *
+   * <p>All tracks must be added before {@linkplain #writeSampleData writing any samples}.
+   *
+   * @param format The {@link Format} of the track.
+   * @return A track id for this track, which should be passed to {@link #writeSampleData}.
+   */
   public int addTrack(Format format) {
     Track track = fragmentedMp4Writer.addTrack(/* sortKey= */ 1, format);
     trackIdToTrack.append(track.id, track);
@@ -186,7 +193,7 @@ public final class FragmentedMp4Muxer implements Muxer {
   }
 
   /**
-   * {@inheritDoc}
+   * Writes encoded sample data.
    *
    * <p>Samples are written to the disk in batches. If {@link
    * Builder#setSampleCopyingEnabled(boolean) sample copying} is disabled, the {@code byteBuffer}
@@ -202,7 +209,6 @@ public final class FragmentedMp4Muxer implements Muxer {
    * @param bufferInfo The {@link BufferInfo} related to this sample.
    * @throws MuxerException If there is any error while writing data to the disk.
    */
-  @Override
   public void writeSampleData(int trackId, ByteBuffer byteBuffer, BufferInfo bufferInfo)
       throws MuxerException {
     try {
@@ -218,7 +224,7 @@ public final class FragmentedMp4Muxer implements Muxer {
   }
 
   /**
-   * {@inheritDoc}
+   * Adds {@linkplain Metadata.Entry metadata} about the output file.
    *
    * <p>List of supported {@linkplain Metadata.Entry metadata entries}:
    *
@@ -236,13 +242,18 @@ public final class FragmentedMp4Muxer implements Muxer {
    *     IllegalArgumentException} is thrown if the {@linkplain Metadata.Entry metadata} is not
    *     supported.
    */
-  @Override
   public void addMetadataEntry(Metadata.Entry metadataEntry) {
     checkArgument(MuxerUtil.isMetadataSupported(metadataEntry), "Unsupported metadata");
     metadataCollector.addMetadata(metadataEntry);
   }
 
-  @Override
+  /**
+   * Closes the file.
+   *
+   * <p>The muxer cannot be used anymore once this method returns.
+   *
+   * @throws MuxerException If the muxer fails to finish writing the output.
+   */
   public void close() throws MuxerException {
     try {
       fragmentedMp4Writer.close();