+++ RELEASENOTES.md
@@ -72,6 +72,8 @@
         in both `Mp4Muxer.Builder` and `FragmentedMp4Muxer.Builder`.
     *   `Mp4Muxer.addTrack()` and `FragmentedMp4Muxer.addTrack()` now return an
         `int` track id instead of a `TrackToken`.
+    *   `Mp4Muxer` and `FragmentedMp4Muxer` no longer implement `Muxer`
+        interface.
 *   IMA extension:
 *   Session:
     *   Fix bug where calling a `Player` method on a `MediaController` connected